refactor to client.js
