+++ spec/dummy/config/webpack/test.js
@@ -1,33 +1,11 @@
 process.env.NODE_ENV = process.env.NODE_ENV || 'test'
 
-const environment = require('./environment')
+const clientEnvironment = require('./client')
 const serverConfig = require('./server')
 const merge = require("webpack-merge")
 
 
-//adding exposeLoader
-const exposeLoader = {
-    test: require.resolve('jquery'),
-    use: [ { loader: 'expose-loader', options: 'jQuery' } ]
-}
-environment.loaders.insert('expose', exposeLoader, { after: 'file'} )
-
-//adding es5Loader
-const es5Loader = {
-    test: require.resolve('react'),
-    use: [ { loader: 'imports-loader', options: { shim: 'es5-shim/es5-shim', sham: 'es5-shim/es5-sham' } } ]
-}
-environment.loaders.insert('react', es5Loader, { after: 'sass'} )
-
-//adding jqueryUjsLoader
-const jqueryUjsLoader = {
-    test: require.resolve('jquery-ujs'),
-    use: [ { loader: 'imports-loader', options: { jQuery: 'jquery' } } ]
-}
-environment.loaders.insert('jquery-ujs', jqueryUjsLoader, { after: 'react'} )
-
-
-const clientConfig = merge(environment.toWebpackConfig(), {
+const clientConfig = merge(clientEnvironment.toWebpackConfig(), {
     mode: 'development',
     entry: {
         'vendor-bundle': [
@@ -37,7 +15,7 @@ const clientConfig = merge(environment.toWebpackConfig(), {
     output: {
         filename: '[name].js',
         chunkFilename: '[name].bundle.js',
-        path: environment.config.output.path
+        path: clientEnvironment.config.output.path
     }
 })
